requirements.txt
kafka-python
openai
posthog
feedparser
newsapi-python
python-dotenv
apscheduler
